# doc/openapi.yaml
openapi: 3.0.0
info:
    version: 1.0.0
    title: incwadi API
    description: incwadi is a book database to manage your books.
...
servers:
    - url: http://localhost:8000/api
...
paths:
...
    /public/book/{id}:
        get:
            tags:
                - public
            description: Shows a Book
            parameters:
                - name: id
                  in: path
                  description: ID of the book
                  schema:
                      type: string
                  required: true
            responses:
                "200":
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PublicBookResponse"
                "404":
                    description: Branch not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...
    /bookmark/:
        get:
            tags:
                - bookmark
            description: List all bookmarks for a branch
            responses:
                "200":
                    description: Success
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/BookmarkResponse"
                "401":
                    description: Access token is missing or invalid
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /bookmark/new:
        post:
            tags:
                - bookmark
            description: Create a new bookmark
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/BookmarkRequest"
            responses:
                "200":
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BookmarkResponse"
                "401":
                    description: Access token is missing or invalid
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /bookmark/{id}:
        get:
            tags:
                - bookmark
            parameters:
                - name: id
                  in: path
                  required: true
                  description: Id of the bookmark
                  schema:
                    type: string
            description: Shows a bookmark
            responses:
                "200":
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BookmarkResponse"
                "401":
                    description: Access token is missing or invalid
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Not Found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
        put:
            tags:
                - bookmark
            parameters:
                - name: id
                  in: path
                  required: true
                  description: Id of the bookmark
                  schema:
                    type: string
            description: Edit a bookmark
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/BookmarkRequest"
            responses:
                "200":
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BookmarkResponse"
                "401":
                    description: Access token is missing or invalid
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Not Found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
        delete:
            tags:
                - bookmark
            parameters:
                - name: id
                  in: path
                  required: true
                  description: Id of the bookmark
                  schema:
                    type: string
            description: Delete a bookmark
            responses:
                "200":
                    description: Success
                    content:
                        application/json:
                            schema:
                                type: object
                                required:
                                    - msg
                                properties:
                                    msg:
                                        type: string
                "401":
                    description: Access token is missing or invalid
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Not Found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        PublicBookResponse:
            type: object
            required:
                - books
                - counter
            properties:
                books:
                    type: array
                    items:
                        type: object
                        required:
                            - id
                            - currency
                            - title
                            - shortDescription
                            - authorFirstname
                            - authorSurname
                            - genre
                            - price
                            - releaseYear
                            - type
                            - branchName
                            - branchOrdering
                            - cond
                            - cover_s
                            - cover_m
                            - cover_l
                            - format
                        properties:
                            id:
                                type: string
                            currency:
                                type: string
                            title:
                                type: string
                            shortDescription:
                                type: string
                            authorFirstname:
                                type: string
                            authorSurname:
                                type: string
                            genre:
                                type: string
                            price:
                                type: integer
                            releaseYear:
                                type: integer
                            type:
                                type: string
                            branchName:
                                type: string
                            branchOrdering:
                                type: string
                            cond:
                                type: string
                            cover_s:
                                type: string
                            cover_m:
                                type: string
                            cover_l:
                                type: string
                            format:
                                type: string
                counter:
                    type: integer
...
        BookmarkRequest:
            type: object
            required:
                - url
                - name
            properties:
                url:
                    type: string
                name:
                    type: string
        BookmarkResponse:
            type: object
            required:
                - id
                - url
                - name
            properties:
                id:
                    type: string
                url:
                    type: string
                name:
                    type: string
        ErrorResponse:
            type: object
            required:
                - code
                - message
            properties:
                code:
                    type: integer
                message:
                    type: string
security:
    - bearerAuth: []